requirements.txt
Flask
dnspython
openpyxl
smtplib
